astrology_ai/requirements.txt
# Core dependencies for Astrology AI Phase 1
PyPDF2>=3.0.1
pdfplumber>=0.10.0
spacy>=3.6.0
click>=8.1.0
pydantic>=2.0.0
sqlalchemy>=2.0.0

# Configuration management
PyYAML>=6.0

# Additional utilities
python-dateutil>=2.8.2